add GenericETHTxEntity
