+++ app/src/main/java/com/keystone/cold/db/entity/TxEntity.java
@@ -28,7 +28,7 @@ import com.keystone.cold.model.Tx;
 
 @Entity(tableName = "txs", indices = {@Index("txId")})
 
-public class TxEntity implements Tx, FilterableItem {
+public class TxEntity implements Tx {
 
     @PrimaryKey
     @NonNull
